chore: move directories
