+++ src/Code/CodeViewer/index.tsx
@@ -1,8 +1,7 @@
 import * as cn from 'classnames';
-import 'prismjs';
 import * as React from 'react';
 
-import { Classes } from '../classes';
+import { Classes } from '../../classes';
 import { astToReact } from './utils/astToReact';
 import { parseCode } from './utils/parseCode';
