+++ src/Code/CodeViewer/utils/__tests__/fixtures/javascript.json
